+++ libmei/atts_visual.h
@@ -960,14 +960,14 @@ public:
      * to the default value)
      **/
     ///@{
-    void SetForm(meterSigVis_FORM form_) { m_form = form_; }
-    meterSigVis_FORM GetForm() const { return m_form; }
+    void SetForm(data_METERFORM form_) { m_form = form_; }
+    data_METERFORM GetForm() const { return m_form; }
     bool HasForm() const;
     ///@}
 
 private:
     /** Indicates to what degree the harmonic label is supported by the notation. **/
-    meterSigVis_FORM m_form;
+    data_METERFORM m_form;
 
     /* include <attform> */
 };
@@ -996,8 +996,8 @@ public:
      * to the default value)
      **/
     ///@{
-    void SetMeterForm(meterSigDefaultVis_METERFORM meterForm_) { m_meterForm = meterForm_; }
-    meterSigDefaultVis_METERFORM GetMeterForm() const { return m_meterForm; }
+    void SetMeterForm(data_METERFORM meterForm_) { m_meterForm = meterForm_; }
+    data_METERFORM GetMeterForm() const { return m_meterForm; }
     bool HasMeterForm() const;
     //
     void SetMeterShowchange(data_BOOLEAN meterShowchange_) { m_meterShowchange = meterShowchange_; }
@@ -1007,7 +1007,7 @@ public:
 
 private:
     /** Contains an indication of how the meter signature should be rendered. **/
-    meterSigDefaultVis_METERFORM m_meterForm;
+    data_METERFORM m_meterForm;
     /**
      * Determines whether the old meter signature should be displayed when the meter
      * signature changes.